Removed old ebuilds, Added new one (0.6) with (untested!) EPREFIX compatibility

+++ source/app-admin/admin-tools-0.6/adduse.py
@@ -6,50 +6,52 @@ Copyright Information
 2010 Johann Schmitz, Duesseldorf, Germany
 
 @author Johann Schmitz (johann@j-schmitz.net)
-@version $Id$
+@version $Id: adduse.py 1401 2010-11-12 18:53:35Z johann $
 
 Web:  www.j-schmitz.net
 Mail: johann@j-schmitz.net
-
-Created: 2010-01-24
 '''
 
 import os
 import shutil
 import sys
 import re
+import portage
 
 comment_re = re.compile("^\s*#.*", re.IGNORECASE)
 uses_re = re.compile("^\s*([\w_\-0-9\+]+/[\w_\-0-9\+]+) (.*)", re.IGNORECASE)
 split_re = re.compile("\s+", re.IGNORECASE)
 
-PACKAGE_USE="/etc/portage/package.use"
+prefix_root = portage.settings['EPREFIX']
+PACKAGE_USE=os.path.join(prefix_root, "/etc/portage/package.use")
 
 def read_uses():
 	packages = {}
 	garbage = []
-	with open(PACKAGE_USE) as f:
-		for l in f:
-			line = l.strip() 
-			if line != "" and not comment_re.match(line):
-				match = uses_re.match(line)
-				if match:
-					pkg = match.group(1)
-					uses = split_re.split(match.group(2))
-					#print "%s: %s" % (pkg, ' '.join(uses))
-
-					if not pkg or len(pkg.strip()) == 0 or not uses or len(uses) == 0:
-						print "Skipping garbage: %s" % line
-						continue 
-
-					if pkg in packages.keys():
-						u = packages[pkg]
-						u.extend(uses)
-						packages[pkg] = u
+	
+	if os.path.exists(PACKAGE_USE):
+		with open(PACKAGE_USE) as f:
+			for l in f:
+				line = l.strip() 
+				if line != "" and not comment_re.match(line):
+					match = uses_re.match(line)
+					if match:
+						pkg = match.group(1)
+						uses = split_re.split(match.group(2))
+						#print "%s: %s" % (pkg, ' '.join(uses))
+	
+						if not pkg or len(pkg.strip()) == 0 or not uses or len(uses) == 0:
+							print "Skipping garbage: %s" % line
+							continue 
+	
+						if pkg in packages.keys():
+							u = packages[pkg]
+							u.extend(uses)
+							packages[pkg] = u
+						else:
+							packages[pkg] = uses
 					else:
-						packages[pkg] = uses
-				else:
-					garbage.append(line)
+						garbage.append(line)
 	return packages, garbage
 
 def negate(use):
@@ -95,7 +97,8 @@ def add_use(pkg, uses):
 
 	uses = clean_uses(current_uses)
 
-	shutil.copyfile(PACKAGE_USE, PACKAGE_USE + ".backup")
+	if os.path.exists(PACKAGE_USE):
+		shutil.copyfile(PACKAGE_USE, PACKAGE_USE + ".backup")
 
 	with open(PACKAGE_USE, 'w') as f:
 		#for k, v in uses.iteritems():
@@ -106,12 +109,19 @@ def add_use(pkg, uses):
 		for x in trash:
 			f.write("%s\n" % x)
 
-
 if __name__ == '__main__':
 	args = sys.argv[1:]
 
-	if os.getuid() != 0:
-		print "Please run this script as root."
+	if not os.path.exists(PACKAGE_USE):
+		if not os.path.exists(os.path.dirname(PACKAGE_USE)):
+			try:
+				os.makedirs(os.path.dirname(PACKAGE_USE), 0755)
+			except:
+				print "Could not create directory %s!" % os.path.dirname(PACKAGE_USE)
+				sys.exit(1)
+
+	if os.path.exists(PACKAGE_USE) and not os.access(PACKAGE_USE, os.W_OK):
+		print "%s is not writable!" % PACKAGE_USE
 		sys.exit(1)
 
 	if len(args) >= 2:
@@ -119,3 +129,4 @@ if __name__ == '__main__':
 	else:
 		print "USAGE: adduse <package> use1... useN"
 		sys.exit(1)
+
